Add concept of task executor to reuse common code

+++ payload/recipes/import_dump.bash
@@ -33,9 +33,11 @@ zcat $DB_FILE_PATH | $MYSQL $DATABASE_NAME
 
 echo "Database imported."
 
-if [ -d "ezplatform/web/var" ]; then
-    rm -rf ezplatform/web/var
-fi
+if [ -f $STORAGE_FILE_PATH ]; then
+    if [ -d "ezplatform/web/var" ]; then
+        rm -rf ezplatform/web/var
+    fi
 
-tar xvzf $STORAGE_FILE_PATH -C ezplatform/web/
-echo "Storage imported."
+    tar xvzf $STORAGE_FILE_PATH -C ezplatform/web/
+    echo "Storage imported."
+fi
